feat: add more models to ObjectBox
