+++ app/objectbox-models/default.json
@@ -28,10 +28,119 @@
         }
       ],
       "relations": []
+    },
+    {
+      "id": "2:8483416042051247361",
+      "lastPropertyId": "3:6571976301610582466",
+      "name": "Degree",
+      "properties": [
+        {
+          "id": "1:8250887022207590829",
+          "name": "id",
+          "type": 6,
+          "flags": 1
+        },
+        {
+          "id": "2:7503190847987138374",
+          "name": "hash",
+          "indexId": "2:8009933817647220507",
+          "type": 9,
+          "flags": 34848
+        },
+        {
+          "id": "3:6571976301610582466",
+          "name": "name",
+          "type": 9
+        }
+      ],
+      "relations": []
+    },
+    {
+      "id": "3:5587073524138502032",
+      "lastPropertyId": "3:4392348489214179879",
+      "name": "Speciality",
+      "properties": [
+        {
+          "id": "1:8685070690975252687",
+          "name": "id",
+          "type": 6,
+          "flags": 1
+        },
+        {
+          "id": "2:4353905416613908921",
+          "name": "hash",
+          "indexId": "3:6717844509497428362",
+          "type": 9,
+          "flags": 34848
+        },
+        {
+          "id": "3:4392348489214179879",
+          "name": "name",
+          "type": 9
+        }
+      ],
+      "relations": []
+    },
+    {
+      "id": "4:1325344862223154709",
+      "lastPropertyId": "4:2100906686007014976",
+      "name": "Subject",
+      "properties": [
+        {
+          "id": "1:1537038551161963375",
+          "name": "id",
+          "type": 6,
+          "flags": 1
+        },
+        {
+          "id": "2:6750811827685143406",
+          "name": "hash",
+          "indexId": "4:5698629409340133043",
+          "type": 9,
+          "flags": 34848
+        },
+        {
+          "id": "3:1859499814414271886",
+          "name": "name",
+          "type": 9
+        },
+        {
+          "id": "4:2100906686007014976",
+          "name": "shortName",
+          "type": 9
+        }
+      ],
+      "relations": []
+    },
+    {
+      "id": "5:8975527960852969049",
+      "lastPropertyId": "3:1930945219455504960",
+      "name": "Title",
+      "properties": [
+        {
+          "id": "1:1174589416841611618",
+          "name": "id",
+          "type": 6,
+          "flags": 1
+        },
+        {
+          "id": "2:3296155943870756019",
+          "name": "hash",
+          "indexId": "5:1171212742400694121",
+          "type": 9,
+          "flags": 34848
+        },
+        {
+          "id": "3:1930945219455504960",
+          "name": "name",
+          "type": 9
+        }
+      ],
+      "relations": []
     }
   ],
-  "lastEntityId": "1:506341774016937552",
-  "lastIndexId": "1:4079983433625419727",
+  "lastEntityId": "5:8975527960852969049",
+  "lastIndexId": "5:1171212742400694121",
   "lastRelationId": "0:0",
   "lastSequenceId": "0:0",
   "modelVersion": 5,
