+++ app/src/test/java/net/wvffle/android/pb/schedule/BackendApiTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 
 import net.wvffle.android.pb.schedule.api.BackendApi;
-import net.wvffle.android.pb.schedule.api.syncedcollectionentry.entries.Room;
+import net.wvffle.android.pb.schedule.api.db.models.Room;
 
 public class BackendApiTest {
     @Test
